doxygen comments & refactor

+++ slcan_future.h
@@ -8,13 +8,18 @@
 #include "slcan_err.h"
 
 
-#define SLCAN_FUTURE_RESULT(RES) ((void*)(RES))
+//! Преобразует значение к результату будущего.
+#define SLCAN_FUTURE_RESULT(RES) ((void*)(long)(RES))
 
+//! Преобразует результат будущего в прочие типы.
 #define SLCAN_FUTURE_RESULT_CAST(T, RES) ((T)(RES))
 #define SLCAN_FUTURE_RESULT_CAST_LONG(RES) SLCAN_FUTURE_RESULT_CAST(long, (RES))
 #define SLCAN_FUTURE_RESULT_CAST_ULONG(RES) SLCAN_FUTURE_RESULT_CAST(unsigned long, (RES))
+//! Преобразует результат будущего в int.
 #define SLCAN_FUTURE_RESULT_INT(RES) ((int)SLCAN_FUTURE_RESULT_CAST_LONG(RES))
+//! Преобразует результат будущего в unsigned int.
 #define SLCAN_FUTURE_RESULT_UINT(RES) ((unsigned int)SLCAN_FUTURE_RESULT_CAST_LONG(RES))
+//! Преобразует результат будущего в slcan_err_t.
 #define SLCAN_FUTURE_RESULT_ERR(RES) ((slcan_err_t)SLCAN_FUTURE_RESULT_CAST_LONG(RES))
